Add tensorflow-gpu configuration to configure.sh.

+++ configure.sh
@@ -21,13 +21,57 @@ function write_action_env_to_bazelrc() {
   write_to_bazelrc "build --action_env $1=\"$2\""
 }
 
-rm .bazelrc
-if python -c "import tensorflow" &> /dev/null; then
-    echo 'using installed tensorflow'
-else
+# Remove .bazelrc if it already exist
+[ -e .bazelrc ] && rm .bazelrc
+
+# Check if we are building GPU or CPU ops, default CPU
+while [[ "$INSTALL_CPU" == "" ]]; do
+  read -p "Do you want to build ops again TensorFlow CPU pip package?"\
+" Y or enter for CPU (tensorflow), N for GPU (tensorflow-gpu). [Y/n] " INPUT
+  case $INPUT in
+    [Yy]* ) echo "Build with CPU pip package."; INSTALL_CPU=1;;
+    [Nn]* ) echo "Build with GPU pip package."; INSTALL_CPU=0;;
+    "" ) echo "Build with CPU pip package."; INSTALL_CPU=1;;
+    * ) echo "Invalid selection: " $INPUT;;
+  esac
+done
+
+
+
+# CPU
+if [[ "$INSTALL_CPU" == "1" ]]; then
+
+  # Check if it's installed
+  if [[ $(pip show tensorflow) == *tensorflow* ]]; then
+    echo 'Using installed tensorflow'
+  else
+    # Uninstall GPU version if it is installed.
+    if [[ $(pip show tensorflow-gpu) == *tensorflow-gpu* ]]; then
+      echo 'Already have tensorflow-gpu installed. Uninstalling......\n'
+      pip uninstall tensorflow-gpu
+    fi
+    # Install CPU version
+    echo 'Installing tensorflow......\n'
     pip install tensorflow
-    pip install tensorflow_estimator
+  fi
+
+else
+
+  # Check if it's installed
+  if [[ $(pip show tensorflow-gpu) == *tensorflow-gpu* ]]; then
+    echo 'Using installed tensorflow-gpu'
+  else
+    # Uninstall CPU version if it is installed.
+    if [[ $(pip show tensorflow) == *tensorflow* ]]; then
+      echo 'Already have tensorflow non-gpu installed. Uninstalling......\n'
+      pip uninstall tensorflow
+    fi
+    # Install CPU version
+    echo 'Installing tensorflow-gpu .....\n'
+    pip install tensorflow-gpu
+  fi
 fi
+
 
 TF_CFLAGS=( $(python -c 'import tensorflow as tf; print(" ".join(tf.sysconfig.get_compile_flags()))') )
 TF_LFLAGS=( $(python -c 'import tensorflow as tf; print(" ".join(tf.sysconfig.get_link_flags()))') )
